// apps/etterlatte-node-server/omstillingsstoenad-ui/src/api/dto/FellesOpplysninger.ts
export interface Opplysning<T> {
    spoersmaal: String
    svar: T
}

export interface BetingetOpplysning<T, R> {
    spoersmaal: String
    svar: T
    opplysning?: R
}

export interface Svar<T> {
    innhold: T
}

export interface FritekstSvar extends Svar<String> {
    innhold: String
}

export interface DatoSvar extends Svar<Date | String> {
    innhold: Date | String
}

export interface EnumSvar<E> extends Svar<String> {
    innhold: String
    verdi: E
}

export enum JaNeiVetIkke {
    JA = 'JA',
    NEI = 'NEI',
    VET_IKKE = 'VET_IKKE',
}

export interface Utenlandsadresse {
    land: Opplysning<FritekstSvar>
    adresse: Opplysning<FritekstSvar>
}

export interface UtbetalingsInformasjon {
    kontonummer?: Opplysning<FritekstSvar>
    utenlandskBankNavn?: Opplysning<FritekstSvar>
    utenlandskBankAdresse?: Opplysning<FritekstSvar>
    iban?: Opplysning<FritekstSvar>
    swift?: Opplysning<FritekstSvar>
    skattetrekk?: BetingetOpplysning<EnumSvar<JaNeiVetIkke>, Opplysning<FritekstSvar> | undefined>
}

export interface Kontaktinfo {
    telefonnummer: Opplysning<FritekstSvar>
}

export enum BankkontoType {
    NORSK = 'NORSK',
    UTENLANDSK = 'UTENLANDSK',
}